# docs/swagger.yaml
basePath: /
definitions:
...
  websocket.ErrorResponse:
    properties:
      code:
        example: 400
        type: integer
      message:
        example: Invalid request
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Realtime chat rooms with WebSocket and REST
  title: Chatters API
  version: 0.1.0
paths:
...
  /ws/{room_id}:
    get:
      description: Opens a WebSocket connection to the specified room. Optionally
        provide a username.
      parameters:
      - description: Room ID (1-999999999)
        in: path
        name: room_id
        required: true
        type: integer
      - description: Username for chat. If omitted, 'Anonymous' is used
        in: query
        name: username
        type: string
      - description: Room password if required
        in: query
        name: password
        type: string
      - description: Host token for room management privileges
        in: query
        name: host_token
        type: string
      responses:
        "101":
          description: Switching Protocols (WebSocket upgraded)
          schema:
            type: string
        "400":
          description: Bad request or validation error
          schema:
            $ref: '#/definitions/websocket.ErrorResponse'
        "401":
          description: Unauthorized - invalid password or host token
          schema:
            $ref: '#/definitions/websocket.ErrorResponse'
        "404":
          description: Room not found
          schema:
            $ref: '#/definitions/websocket.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/websocket.ErrorResponse'
      summary: Connect to WebSocket room
      tags:
      - websocket
swagger: "2.0"